+++ apps/block_scout_web/lib/block_scout_web/controllers/api/rpc/eth_controller.ex
@@ -29,6 +29,8 @@ defmodule BlockScoutWeb.API.RPC.EthController do
     3 => "fourth"
   }
 
+  def methods, do: @methods
+
   def eth_request(%{body_params: %{"_json" => requests}} = conn, _) when is_list(requests) do
     responses = responses(requests)
 
@@ -106,7 +108,11 @@ defmodule BlockScoutWeb.API.RPC.EthController do
   end
 
   defp render_log(log) do
-    topics = Enum.reject([log.first_topic, log.second_topic, log.third_topic, log.fourth_topic], &is_nil/1)
+    topics =
+      Enum.reject(
+        [log.first_topic, log.second_topic, log.third_topic, log.fourth_topic],
+        &is_nil/1
+      )
 
     %{
       "address" => to_string(log.address_hash),
@@ -245,7 +251,8 @@ defmodule BlockScoutWeb.API.RPC.EthController do
   defp to_block_numbers(from_block, to_block, max_block_number, pending_block_number) do
     actual_pending_block_number = pending_block_number || max_block_number
 
-    with {:ok, from} <- to_block_number(from_block, max_block_number, actual_pending_block_number),
+    with {:ok, from} <-
+           to_block_number(from_block, max_block_number, actual_pending_block_number),
          {:ok, to} <- to_block_number(to_block, max_block_number, actual_pending_block_number) do
       {:ok, from, to}
     end